added Route DAO and implementation class fixed connection problem

+++ EpamFinalProject/src/main/java/com/example/epamfinalproject/Entities/User.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class User {
-    private int id;
+    private long id;
     private String firstName;
     private String lastName;
     private String login;
@@ -26,7 +26,7 @@ public class User {
     }
 
     public static class UserBuilder {
-        private int id;
+        private long id;
         private String firstName;
         private String lastName;
         private String login;
@@ -35,7 +35,7 @@ public class User {
 
         public UserBuilder() {}
 
-        public UserBuilder id(int id) {
+        public UserBuilder id(long id) {
             this.id = id;
             return this;
         }
